receiver/pyproject.toml
[tool.poetry]
name = "userbot-receiver"
version = "0.1.0"
description = "Telegram user account receiver service"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyrogram = "^2.0"
tgcrypto = "^1.2.5"
Flask = "^2.3"
python-dotenv = "^1.0"
requests = "^2.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
